Fix mongodb replicaset recipe to include node establishing set
Mongodb complained when the node attempted to initiate a replicaset connection
because the node had not included itself among the replicaset members.  This
occurs because the mongodb definition creates the members array by doing a
knife search on all nodes with a matching the replicaset name, but since the
chef-run is not completed at that point, the current node is not included.