Remove check if mongo gem needs to be added and just add it in the recipes
Running the gem install multiple times will not hurt anything.
Right now checking to see if the recipe is included is a bit backwards and causes issues if the cookbook is being wrapped.
Unless the cookbook is directly in the nodes run_list, it will not show up as included.