Remove ip to hostname conversion in replica-set function
When configuring the mongodb replicat set, one of the steps was to
force the usage of `fqdn` address instead of ips in the member
configuration.

This happens to cause erros when you provide the `use_ip_address`
option, that is needed in cases were the machines `fqdn` is not a valid
DNS record, making the replica set members unreachable.

I removed the code that make the conversion, assuming that this is
acctually a bad behaviour for the cookbook.