Fix some packages not upgrading in Debian
Only mongodb-org and mongodb-org-server are upgrading, the rest are left in current state. This is because when installing multiple packages we need to specify version for each one of them using array notation, otherwise only the first package is being checked against specified version number. In this case it was mongodb-org-server, which is why it was successfully upgraded. Ultimately, it's easier to just run install for each package individually, which is what this change does.

Reference: https://docs.chef.io/resource_package.html#multiple-packages.